Migrate::공유된 세트 컴포넌트 타입스크립트로 변환
